pyproject.toml
[tool.poetry]
name = "km-server"
version = "0.1.0"
description = ""
authors = ["monkjiro <thenullmastermind@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
